Add important Java-8-compat casts.
The last release, 0.4.1, won't work if using Java 8 because some part of the build process used Java 9 or higher, and each of the ColorfulBatch classes uses a Buffer method that was effectively removed from ShortBuffer in Java 9. Casting each ShortBuffer to a Buffer and then calling position(int) or limit(int) on it seems to work.