Huge effort for slight color adjustment.
The colors "sky" and "denim" in the simple palette were at something of an anomalous position in the color space, so some common colors (like the blue-green that the ocean can be) matched to an oddly-purple version of denim, instead of a bluish-green sky color (which was probably too light to consider). It only took some small changes to each to get DescriptionMatcher looking better (denim is more grayish instead of purplish, while sky is more saturated and slightly darker).


Former-commit-id: b02e58beaf46f32767d676bfa78745182754a259