Fix the strangest GWT alpha bug...
ColorTools.toEditedFloat() (in the rgb, cielab, and oklab packages) seemed to be rather imprecise and caused alpha to go below the threshold of 254 need for this to consider it opaque. This never happened on desktop... It likely has to do with a specific constant being too small for a GWT "float-like" value to store accurately.


Former-commit-id: e1e918fba45c8380401e26cb957d33858b6918b5