The lightness-exponent approach wasn't great.
I went back to using a lerp towards white or black, which is sometimes the same as lighten or darken. This seems to maintain the visibility of a lightness change for most colors. I also reordered lightness and saturation changes again; because a lerp with black or white also desaturates, having rich or dull applied after that helps. I only did Oklab and RGB here; more are coming.